refactor(game-menu): split UI into focused modules
