+++ app/src/test/java/com/limelight/binding/audio/AudioHapticsRuntimePolicyTest.kt
@@ -1,4 +1,4 @@
-package com.limelight
+package com.limelight.binding.audio
 
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
